Filter assetUrl props from component, and replace orignal prop to make usage clearer
