+++ src/behaviours/index.js
@@ -3,4 +3,4 @@ export { default as scrollable } from './scrollable';
 export { default as autoInitialisedForm } from './autoInitialisedForm';
 export { default as withBounds } from './withBounds';
 export { default as withOptionsFromSelector } from './withOptionsFromSelector';
-export { default as pathAsAssetUrl } from './pathAsAssetUrl';
+export { default as injectAssetUrl } from './injectAssetUrl';
